drm/edid: Don't send bogus aspect ratios in AVI infoframes
If the user mode would specify an aspect ratio other than 4:3 or 16:9
we now silently ignore it. Maybe a better apporoach is to return an
error? Let's try that.

Also toss in a warn to the infoframe packing code to let the developer
know they're trying to pass in an unsupported aspect ratio.

Cc: Shashank Sharma <shashank.sharma@intel.com>
Cc: Sean Paul <seanpaul@chromium.org>
Cc: Jose Abreu <Jose.Abreu@synopsys.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Emil Velikov <emil.l.velikov@gmail.com>
Cc: Thierry Reding <thierry.reding@avionic-design.de>
Fixes: a68362fe3e84 ("drm: Add and handle new aspect ratios in DRM layer")
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>