drm/i915: Clear plls from the new crtc state
In order to retry the DPLL selection we must make
intel_modeset_clear_plls() clear the DPLL from the new state rather than
the old state. This way the tracking remains correct even if we
reattempt the DPLL selection multiple times. This works just fine the
first time around as well since we don't clear the DPLL selection from
the crtc state when we duplicate it.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>