dim: support git worktree for aux checkouts
If available by default. This saves quite a pile of disk-space that
imo making it the default is justified.

Note that this will break the rebuild-nightly script for now,
follow-up patches will fix that.

v2: Rebase to put the dim rework at the end of the series.

v3: s/remote_url/url/

Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>