drm/i915: Workaround VLV/CHV sprite1->sprite0 enable underrun
On VLV/CHV enabling sprite0 when sprite1 has already been enabled may
lead to an underrun. This only happens when sprite0 FIFO size is zero
prior to enabling it. Hence an effective workaround is to always
allocate at least one cacheline for sprite0 when sprite1 is active.

I've not observed this sort of failure during any other type of plane
enable/disable sequence.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>