drm/i915: Clean up the HWSTAM mess
Currently we're unmasking some random looking bits in HWSTAM
on gen3/4/5. The two bits we apparently unmask are 0 and 12,
and also bits 16-31 on gen4/5.
What those bits do depends on the gen as follows:
 bit 0: Breakpoint (gen2), ASLE (gen3), reserved (gen4/5)
 bit 12: Sync flush statusa
 bit 16-31: The ones that can unmasked seem to be mostly some
            display stuff. Bit 18 is the PIPE_CONTROL notify,
	    which might be the only intresting one.

Bit 0 at least looks quite bogus here. Bit 12 might have some meaning
as bspec tells us:
"This bit is toggled when the Instruction Parser completes a flush with the sync enable bit
 active in the INSTPM register. The toggle event will happen after all the graphics engines
 are flushed. The HW Status DWord write resulting from this toggle will cause the CPU’s
 view of graphics memory to be coherent as well (flush and invalidate the render cache)."

So I don't know whether we actually depend on that status page write
somehow. Extra seqno coherency by accident perhaps? Except we don't
even unmask the user interrupt bit in HWSTAM, and sync flush isn't
something we use normally, so seems unlikely. So let's just assume
we don't need any of this and mask everything in HWSTAM.

From gen6 onwards there's a separate HWSTAM for each engine, and so
we deal with them during the engine setup.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>