drm/i915/sdvo: Implement proper HDMI audio support for SDVO
Our SDVO audio support is pretty bogus. We can't push audio over the
SDVO bus, so trying to enable audio in the SDVO control register doesn't
do anything. In fact it looks like the SDVO encoder will always mix in
the audio coming over HDA, and there's no (at least documented) way to
disable that from our side. So HDMI audio does work currently but only by
luck really. What is missing though is the ELD.

To pass the ELD to the audio driver we need to write it to magic buffer
in the SDVO encoder hardware which then gets pulled out via HDA in the
other end. Ie. pretty much the same thing we had for native HDMI before
we started to just pass the ELD between the drivers. This sort of
explains why we even have that silly hardware buffer with native HDMI.

To make sure we actually have a proper ELD we must remember to call
drm_edid_to_eld() in .get_modes(), but in fact we can just switch over
to using intel_connector_update_modes() which takes care of everything
for us.

This also fixes our state readout since we can now query the SDVO
encoder about the state of the "ELD valid" and "presence detect"
bits. As mentioned those don't actually control whether audio gets
sent over the HDMI cable, but it's the best we can do.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>