drm/i915: Add windowing for primary planes on gen2/3 and chv
Plane B and C (note that we don't actually expose plane C currently)
on gen2/3 have a window generator, as does the primary plane on CHV
pipe B. So let's allow positioning of these planes freely within the
pipe source area.

Plane A on gen2/3 seems to have some kind of partial
window generator which would perhaps allow you to cut the plane off
midway through the scanout, but it would still have to start at the top
and be full width. That's doesn't sound all that useful, so for
simplicity let's just keep to the idea that plane A has to be
fullscreen.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>