drm/i915/sdvo: Try to send the SPD infoframe
We should be able to use the general purpose HDMI buffers in SDVO to
send other infoframes besides AVI. Let's try to send the SPD.

For some reason this doesn't actually work on my SiI1392. There are
no errors from the SDVO encoder, but the display doesn't show the SPD
information. Not sure what's going on since I don't have anything that
could analyze the infoframes coming off the wire. I supposes it's
possible that this is just a problem specific to the SiI1392. AVI
infoframe and ELD would appear to work just fine.

This also fixes our state readout since we can now query the SDVO
encoder about the state of the "ELD valid" and "presence detect"
bits. As mentioned those don't actually control whether audio gets
sent over the HDMI cable, but it's the best we can do.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>