drm/i915: Extract skl_universal_plane_init()
There's not much point in following the primary vs. sprite
for the SKL+ universal plane init code. The only difference is
of our own doing in the form of the .check_plane(). Let's make
a small exception for that little detail and otherwise share
the same code to initialize all the universal planes.

Eventually we should eliminate the mess around .check_plane()
as well, but for now let's be happy with some code reduction.

v2: Remember to set up plane->has_fbc
    Make skl_plane_has_ccs() static

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>