drm/i915: Add support for half float framebuffers
gen4+ supports fp16 pixel formats on the primary planes, snb+ on the
sprite planes as well, exception being vlv/chv sprites which don't
support half-float at all.

On ivb fp16 scanout is slightly busted. The output from the plane will
have 1/4 the expected value. For the sprite plane we can fix that
up with the plane gamma unit. For the primary plane we would have to
use the pipe gamma or csc which would affect all the other planes as
well, hence we simply choose not to expose fp16 on the ivb primary
plane. On hsw the primary plane got fixed, but the sprite still has the
problem. So we need to apply the sprite gamma w/a on hsw as well.

On gmch platforms I observed that the plane width must be below 2k
pixels with fp16 or else we get a corrupted image. This limitation
does not seem to be documented in bspec. I verified the exact limit
using the chv pipe B primary plane since it has windowing capability.
The stride limits are unaffected by fp16.

Tested on everyting up to SKL, except SNB which I didn't happen to
have at hand.

v2: Rebase on top of icl fp16
    Deal with HAS_GMCH()

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>