drm/i915: Program plane gamma ramps
All sprite planes have a progammable gamma ramp. Set it up with
a linear ramp on all platforms. This actually matches the reset
value but soon we'll want to reprogram this ramp on some machines,
so let's just set it up across the board.

TODO: At least on vlv/chv the plane gamma only seemed to work
with YUV formats. Is that the case for other gmch platforms as well?
on IVB+ it at least works for half float too. Need to test all
the platforms.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>